+++ fsm/common/src/commonMain/kotlin/dev/inmo/micro_utils/fsm/common/UpdatableStatesMachine.kt
@@ -26,6 +26,12 @@ open class DefaultUpdatableStatesMachine<T : State>(
 ), UpdatableStatesMachine<T> {
     protected val jobsStates = mutableMapOf<Job, T>()
 
+    /**
+     * Realization of this update will use the [Job] of [previousState] in [statesJobs] and [jobsStates] if
+     * [previousState] is [Optional.presented] and [shouldReplaceJob] has returned true for [previousState] and [actualState]. In
+     * other words, [Job] of [previousState] WILL NOT be replaced with the new one if they are "equal". Equality of
+     * states is solved in [shouldReplaceJob] and can be rewritten in subclasses
+     */
     override suspend fun performStateUpdate(previousState: Optional<T>, actualState: T, scope: CoroutineScope) {
         statesJobsMutex.withLock {
             if (compare(previousState, actualState)) {
@@ -52,7 +58,13 @@ open class DefaultUpdatableStatesMachine<T : State>(
         }
     }
 
-    protected open suspend fun compare(previous: Optional<T>, new: T): Boolean = previous.dataOrNull() != new
+    /**
+     * Compare if [previous] potentially lead to the same behaviour with [new]
+     */
+    protected open suspend fun shouldReplaceJob(previous: Optional<T>, new: T): Boolean = previous.dataOrNull() != new
+
+    @Deprecated("Overwrite shouldReplaceJob instead")
+    protected open suspend fun compare(previous: Optional<T>, new: T): Boolean = shouldReplaceJob(previous, new)
 
     override suspend fun updateChain(currentState: T, newState: T) {
         statesManager.update(currentState, newState)